Implement support for VAT
This implements support for handling VAT on invoices, including
automatically calculating the VAT amount as necessary.

Support is also added to the confreg and confsponsor modules to
auto-generate invoiecs with the correct VAT. The membership module is
explicitly excluded from VAT handling, as we know we don't need to
process VAT on that.

Different VAT levels are supported by registering them in the table
VATRate in the Invoices module. This rate can then be specified
individually for each row on an invoice, and for the
confreg/confsponsor modules different rates can be specified for
registrations and sponsorship on a per-conference baseis.

This leads to a number of other changs:

* Invoice amounts now deal in two-decimal fractional values, instead
  of just integers as before.
* This also extends to Adyen and Paypal payments that now handle
  two-decimal fractional values at all integration levels, not just
  at the core API.
* The invoice layout for PGEU has a lot of changes to support showing
  VAT on the different entries as well as in the summary
* Invoice titles are changed in the confreg/confsponsor modules to
  be more consistent
* Sponsor signups now explicitly ask for VAT number

In general, just adding the VAT rates will make the system active, and
not adding it will leave it turned off. To enable the VAT information
colleciton for sponsorships, set the parameter EU_VAT to True in the
local settings.

NOTE! This makes incompatible changes to the PDF invoice API. An
updated implementation for PGEU is included, but downstream
implementations like pgopen need to update the PDF invoice
specification calling parameters (see change to PDFInvoice class)!

+++ postgresqleu/util/payment/adyen.py
@@ -2,6 +2,7 @@ from django.conf import settings
 
 from urllib import urlencode
 from datetime import datetime, timedelta
+from decimal import Decimal
 import hmac
 import hashlib
 import base64
@@ -51,7 +52,7 @@ class _AdyenBase(object):
 		orderdata = "<p>%s</p>" % invoicestr
 		param.update({
 			'merchantReference': '%s%s' % (settings.ADYEN_MERCHANTREF_PREFIX, invoiceid),
-			'paymentAmount': '%s' % (invoiceamount*100,),
+			'paymentAmount': '%s' % (int(invoiceamount*Decimal(100.0)),),
 			'orderData': base64.encodestring(_gzip_string(orderdata.encode('utf-8'))).strip().replace("\n",''),
 			'merchantReturnData': '%s%s' % (settings.ADYEN_MERCHANTREF_PREFIX, invoiceid),
 			'sessionValidity': (datetime.utcnow() + timedelta(minutes=30)).strftime('%Y-%m-%dT%H:%M:%SZ'),
